Always add new LoadResults to a Group
Rather than add items to the load queue directly, the load queue
just keeps a list of Group objects, which in turn keep lists of
LoadResults (including other Groups, which are also LoadResults).

As a result of this change, much of the queueing API moves
to the Group (append, prepend), while the LoadQueue itself
pares down to more load-specific API (load, start).